Login and register

+++ Backend/employee-management/src/main/java/com/rg/employee_management/controller/EmployeeController.java
@@ -1,8 +1,7 @@
 package com.rg.employee_management.controller;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -18,9 +17,11 @@ import com.rg.employee_management.entity.Employees;
 import com.rg.employee_management.exception.ResourceNotFoundException;
 import com.rg.employee_management.repository.employeeRepository;
 
+import jakarta.servlet.http.HttpSession;
+
 @RestController
 @RequestMapping("/api")
-@CrossOrigin("http://localhost:4200")
+@CrossOrigin(origins = "http://localhost:4200")
 public class EmployeeController {
 
     @Autowired
@@ -28,19 +29,35 @@ public class EmployeeController {
 
     // Post Mapping to save new employees to a database.
     @PostMapping("/saveEmployees")
-    public Employees addEmployees(@RequestBody Employees employees) {
-        return repository.save(employees);
+    public ResponseEntity<?> addEmployees(@RequestBody Employees employees, HttpSession session) {
+
+        if(session.getAttribute("user") == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Please login first!");
+        }
+
+        Employees saveEmployees = repository.save(employees);
+
+        return ResponseEntity.ok(saveEmployees);
     }
 
     // Get Mapping to list all the employees of a database.
     @GetMapping("/list")
-    public List<Employees> getAllEmployees() {
-        return repository.findAll();
+    public ResponseEntity<?> getAllEmployees(HttpSession session) {
+        if(session.getAttribute("user") == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Please login first!");
+        }
+
+        return ResponseEntity.ok(repository.findAll());
     }
 
     // Get Mapping to get employee details by Id.
     @GetMapping("/getEmployeeById/{empId}")
-    public ResponseEntity<Employees> getEmployeeById(@PathVariable int empId) {
+    public ResponseEntity<?> getEmployeeById(@PathVariable int empId, HttpSession session) {
+
+        if(session.getAttribute("user") == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Please login first!");
+        }
+
         Employees employees = repository.findById(empId)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee with id not found."));
 
@@ -49,7 +66,12 @@ public class EmployeeController {
 
     // Put Mapping to update employee details.
     @PutMapping("/update/{empId}")
-    public ResponseEntity<Employees> updateEmployee(@PathVariable int empId, @RequestBody Employees employees) {
+    public ResponseEntity<?> updateEmployee(@PathVariable int empId, @RequestBody Employees employees, HttpSession session) {
+
+        if(session.getAttribute("user") == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Please login first!");
+        }
+
         Employees existingEmp = repository.findById(empId)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee not found!!"));
 
@@ -63,11 +85,16 @@ public class EmployeeController {
 
     // Delete Mapping to remove employee from a database.
     @DeleteMapping("/deleteEmp/{empId}")
-    public ResponseEntity<Employees> deleteEmployeeById(@PathVariable int empId) {
+    public ResponseEntity<?> deleteEmployeeById(@PathVariable int empId, HttpSession session) {
+
+        if(session.getAttribute("user") == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Please login first!");
+        }
+
         Employees employees = repository.findById(empId)
                 .orElseThrow(() -> new ResourceNotFoundException("Employee not found!!!"));
         repository.delete(employees);
-        return ResponseEntity.ok(employees);
+        return ResponseEntity.ok("Employee deleted Successfully...");
     }
 
 }
